+++ src/main/java/com/devcourse/checkmoi/domain/post/exception/NotAllowedWriterException.java
@@ -1,7 +1,7 @@
 package com.devcourse.checkmoi.domain.post.exception;
 
 import com.devcourse.checkmoi.global.exception.BusinessException;
-import com.devcourse.checkmoi.global.exception.ErrorMessage;
+import com.devcourse.checkmoi.global.exception.error.ErrorMessage;
 
 public class NotAllowedWriterException extends BusinessException {
