+++ src/main/java/com/devcourse/checkmoi/domain/post/exception/PostNoPermissionException.java
@@ -1,6 +1,6 @@
 package com.devcourse.checkmoi.domain.post.exception;
 
-import com.devcourse.checkmoi.global.exception.ErrorMessage;
+import com.devcourse.checkmoi.global.exception.error.ErrorMessage;
 import com.devcourse.checkmoi.global.exception.InvalidValueException;
 
 public class PostNoPermissionException extends InvalidValueException {
