+++ src/main/java/com/devcourse/checkmoi/domain/post/exception/PostNotFoundException.java
@@ -1,7 +1,7 @@
 package com.devcourse.checkmoi.domain.post.exception;
 
 import com.devcourse.checkmoi.global.exception.EntityNotFoundException;
-import com.devcourse.checkmoi.global.exception.ErrorMessage;
+import com.devcourse.checkmoi.global.exception.error.ErrorMessage;
 
 public class PostNotFoundException extends EntityNotFoundException {
