+++ src/main/java/com/devcourse/checkmoi/domain/study/exception/DuplicateStudyJoinRequestException.java
@@ -1,7 +1,7 @@
 package com.devcourse.checkmoi.domain.study.exception;
 
 import com.devcourse.checkmoi.global.exception.DuplicateException;
-import com.devcourse.checkmoi.global.exception.ErrorMessage;
+import com.devcourse.checkmoi.global.exception.error.ErrorMessage;
 
 public class DuplicateStudyJoinRequestException extends DuplicateException {
