+++ src/main/java/com/devcourse/checkmoi/domain/study/exception/NotAllowedStudyStatusException.java
@@ -1,10 +1,9 @@
 package com.devcourse.checkmoi.domain.study.exception;
 
 import com.devcourse.checkmoi.global.exception.BusinessException;
-import com.devcourse.checkmoi.global.exception.ErrorMessage;
-import com.devcourse.checkmoi.global.exception.InvalidStatusException;
+import com.devcourse.checkmoi.global.exception.error.ErrorMessage;
 
-public class NotAllowedStudyStatusException extends InvalidStatusException {
+public class NotAllowedStudyStatusException extends BusinessException {
 
     public NotAllowedStudyStatusException() {
         super(ErrorMessage.NOT_ALLOWED_STUDY_STATUS);
