+++ src/main/java/com/devcourse/checkmoi/domain/user/exception/UserNoPermissionException.java
@@ -1,6 +1,6 @@
 package com.devcourse.checkmoi.domain.user.exception;
 
-import static com.devcourse.checkmoi.global.exception.ErrorMessage.ACCESS_DENIED;
+import static com.devcourse.checkmoi.global.exception.error.ErrorMessage.ACCESS_DENIED;
 import com.devcourse.checkmoi.global.exception.InvalidValueException;
 
 public class UserNoPermissionException extends InvalidValueException {
