+++ src/main/java/com/devcourse/checkmoi/domain/user/model/UserRole.java
@@ -4,7 +4,8 @@ import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
 public enum UserRole {
-    LOGIN("ROLE_LOGIN");
+    LOGIN("ROLE_LOGIN"),
+    ADMIN("ROLE_ADMIN");
 
     private final String grantedAuthority;
