+++ src/main/java/com/devcourse/checkmoi/global/config/WebConfig.java
@@ -1,18 +1,12 @@
 package com.devcourse.checkmoi.global.config;
 
 import static org.springframework.http.HttpHeaders.LOCATION;
-import com.amazonaws.auth.AWSStaticCredentialsProvider;
-import com.amazonaws.auth.BasicAWSCredentials;
-import com.amazonaws.regions.Regions;
-import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.AmazonS3ClientBuilder;
-import com.devcourse.checkmoi.domain.file.service.S3Upload;
-import com.devcourse.checkmoi.global.config.properties.CorsConfigProperties;
+import com.devcourse.checkmoi.global.config.property.CorsConfigProperties;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @Configuration
@@ -32,23 +26,8 @@ public class WebConfig implements WebMvcConfigurer {
             .exposedHeaders(LOCATION);
     }
 
-    @Bean
-    public AmazonS3 amazonS3Client(AwsConfigProperties awsConfigure) {
-        return AmazonS3ClientBuilder.standard()
-            .withRegion(Regions.fromName(awsConfigure.getRegion()))
-            .withCredentials(
-                new AWSStaticCredentialsProvider(
-                    new BasicAWSCredentials(
-                        awsConfigure.getAccessKey(),
-                        awsConfigure.getSecretKey())
-                )
-            )
-            .build();
+    @Override
+    public void addResourceHandlers(ResourceHandlerRegistry registry) {
+        registry.addResourceHandler("/static/**").addResourceLocations("classpath:/static/");
     }
-
-    @Bean
-    public S3Upload s3Upload(AmazonS3 amazonS3, AwsConfigProperties awsProperties) {
-        return new S3Upload(amazonS3, awsProperties.getUrl(), awsProperties.getBucketName());
-    }
-
 }
