+++ src/main/java/com/devcourse/checkmoi/global/exception/EntityNotFoundException.java
@@ -1,5 +1,7 @@
 package com.devcourse.checkmoi.global.exception;
 
+import com.devcourse.checkmoi.global.exception.error.ErrorMessage;
+
 public class EntityNotFoundException extends BusinessException {
 
     public EntityNotFoundException(String message) {
