+++ src/main/java/com/devcourse/checkmoi/global/exception/InvalidStatusException.java
@@ -1,5 +1,7 @@
 package com.devcourse.checkmoi.global.exception;
 
+import com.devcourse.checkmoi.global.exception.error.ErrorMessage;
+
 public class InvalidStatusException extends BusinessException {
 
     public InvalidStatusException(String message,
@@ -11,8 +13,7 @@ public class InvalidStatusException extends BusinessException {
         super(errorMessage);
     }
 
-    public InvalidStatusException(Throwable cause,
-        ErrorMessage errorMessage) {
+    public InvalidStatusException(Throwable cause, ErrorMessage errorMessage) {
         super(cause, errorMessage);
     }
 }
