+++ src/main/java/com/devcourse/checkmoi/global/exception/InvalidValueException.java
@@ -1,5 +1,7 @@
 package com.devcourse.checkmoi.global.exception;
 
+import com.devcourse.checkmoi.global.exception.error.ErrorMessage;
+
 public class InvalidValueException extends BusinessException {
 
     public InvalidValueException(String value) {
