+++ src/main/java/com/devcourse/checkmoi/global/exception/error/ErrorMessage.java
@@ -1,4 +1,4 @@
-package com.devcourse.checkmoi.global.exception;
+package com.devcourse.checkmoi.global.exception.error;
 
 import org.springframework.http.HttpStatus;
