+++ src/main/java/com/devcourse/checkmoi/global/security/jwt/JwtAuthenticationToken.java
@@ -8,6 +8,7 @@ import org.springframework.security.core.GrantedAuthority;
 public class JwtAuthenticationToken extends AbstractAuthenticationToken {
 
     private final Object principal;
+
     private String credentials;
 
     public JwtAuthenticationToken(Object principal, String credentials) {
