+++ src/main/java/com/devcourse/checkmoi/global/security/jwt/exception/InvalidTokenException.java
@@ -1,6 +1,6 @@
 package com.devcourse.checkmoi.global.security.jwt.exception;
 
-import com.devcourse.checkmoi.global.exception.ErrorMessage;
+import com.devcourse.checkmoi.global.exception.error.ErrorMessage;
 
 public class InvalidTokenException extends TokenException {
