+++ src/main/java/com/devcourse/checkmoi/global/security/jwt/exception/TokenException.java
@@ -1,7 +1,7 @@
 package com.devcourse.checkmoi.global.security.jwt.exception;
 
 import com.devcourse.checkmoi.global.exception.BusinessException;
-import com.devcourse.checkmoi.global.exception.ErrorMessage;
+import com.devcourse.checkmoi.global.exception.error.ErrorMessage;
 
 public class TokenException extends BusinessException {
