+++ src/main/java/com/devcourse/checkmoi/global/security/jwt/util/AuthorizationExtractor.java
@@ -1,4 +1,4 @@
-package com.devcourse.checkmoi.global.security.util;
+package com.devcourse.checkmoi.global.security.jwt.util;
 
 import static org.springframework.http.HttpHeaders.*;
