+++ src/main/resources/application-prod.yml
@@ -41,3 +41,7 @@ jwt:
     expire-length: 3600000
   refresh-token:
     expire-length: 7200000
+
+
+front-url:
+  https://checkmoi.vercel.app
