+++ src/main/resources/application.yml
@@ -12,3 +12,6 @@ spring:
     console:
       enabled: true
       path: /h2-console
+
+front-url:
+  http://localhost:8080
